feat: create isopmorphic media and remove old
