+++ src/Atoms/IsomorphicMedia/index.ts
@@ -1,0 +1,4 @@
+import { IsomorphicMedia, useIsomorphicMedia } from './IsomorphicMedia';
+
+export { useIsomorphicMedia };
+export default IsomorphicMedia;
