--- src/Atoms/LayoutEffectMedia/index.ts
@@ -1,4 +1,0 @@
-import { LayoutEffectMedia, useLayoutEffectMedia } from './LayoutEffectMedia';
-
-export { useLayoutEffectMedia };
-export default LayoutEffectMedia;
